v1.0a
wrote the README.
tested and corrected, now it work (still need to see if the usecase covered  are correct or if we have to make the controls more or less heavy).
made an html with the main tests inside.

NEXT TO DO
make the functions compatible with string template not only id.
make an erase function.